requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic[email]==2.12.5
slowapi>=0.1.8
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
